# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Farewell 2026 Event Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AdminLoginRequest {
  username: string;
  password: string;
}

export interface FacultyLoginRequest {
  name: string;
  password: string;
}

export interface AudienceLoginRequest {
  username: string;
  password: string;
}

export type AuthResponseRole =
  (typeof AuthResponseRole)[keyof typeof AuthResponseRole];

export const AuthResponseRole = {
  admin: "admin",
  faculty: "faculty",
  audience: "audience",
} as const;

export interface AuthResponse {
  role: AuthResponseRole;
  username: string;
  name?: string;
}

export interface Category {
  id: number;
  name: string;
  isActive: boolean;
  activeForVoting: boolean;
}

export interface CreateCategoryRequest {
  name: string;
}

export interface UpdateCategoryRequest {
  name?: string;
  isActive?: boolean;
  activeForVoting?: boolean;
}

export type ParticipantGender =
  (typeof ParticipantGender)[keyof typeof ParticipantGender];

export const ParticipantGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface Participant {
  id: number;
  name: string;
  gender: ParticipantGender;
  categoryId: number;
  categoryName?: string;
  contestantNo?: number;
  photoUrl?: string | null;
  orderIndex: number;
}

export type CreateParticipantRequestGender =
  (typeof CreateParticipantRequestGender)[keyof typeof CreateParticipantRequestGender];

export const CreateParticipantRequestGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface CreateParticipantRequest {
  name: string;
  gender: CreateParticipantRequestGender;
  categoryId: number;
  photoUrl?: string | null;
}

export type UpdateParticipantRequestGender =
  (typeof UpdateParticipantRequestGender)[keyof typeof UpdateParticipantRequestGender];

export const UpdateParticipantRequestGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface UpdateParticipantRequest {
  name?: string;
  gender?: UpdateParticipantRequestGender;
  categoryId?: number;
  photoUrl?: string | null;
}

export interface VotingStatus {
  isOpen: boolean;
  activeCategoryId?: number | null;
  activeCategoryName?: string | null;
}

export interface SetVotingStatusRequest {
  isOpen: boolean;
  activeCategoryId?: number | null;
}

export interface AudienceVoteRequest {
  participantId: number;
  categoryId: number;
}

export interface FacultyScoreRequest {
  participantId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  scoreIntroduction?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  scoreRampwalk?: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  scoreTalent?: number;
}

export interface RampWalkState {
  currentParticipantId: number | null;
  isLive?: boolean;
  ratingActive?: boolean;
  participant?: Participant | null;
}

export interface RampWalkSettingsRequest {
  isLive?: boolean;
  ratingActive?: boolean;
}

export interface NextParticipantRequest {
  participantId: number | null;
}

export interface ParticipantResult {
  participantId: number;
  name: string;
  photoUrl?: string | null;
  categoryId: number;
  categoryName: string;
  contestantNo?: number;
  facultyVotes: number;
  audienceVotes: number;
  totalVotes: number;
  averageRating?: number | null;
  facultyScoreCount: number;
}

export interface ResultsResponse {
  top3: ParticipantResult[];
}

export type ListParticipantsParams = {
  categoryId?: number;
};