# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Farewell 2026 Event Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/faculty/login:
    post:
      operationId: facultyLogin
      tags: [auth]
      summary: Faculty login with name and shared password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacultyLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/audience/login:
    post:
      operationId: audienceLogin
      tags: [auth]
      summary: Audience login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudienceLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /participants:
    get:
      operationId: listParticipants
      tags: [participants]
      summary: List participants
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
    post:
      operationId: createParticipant
      tags: [participants]
      summary: Create a participant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateParticipantRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
  /participants/{id}:
    put:
      operationId: updateParticipant
      tags: [participants]
      summary: Update participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateParticipantRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
    delete:
      operationId: deleteParticipant
      tags: [participants]
      summary: Delete participant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    AdminLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    FacultyLoginRequest:
      type: object
      properties:
        name:
          type: string
        password:
          type: string
      required: [name, password]
    AudienceLoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        role:
          type: string
          enum: [admin, faculty, audience]
        username:
          type: string
        name:
          type: string
      required: [role, username]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        isActive:
          type: boolean
        activeForVoting:
          type: boolean
      required: [id, name, isActive, activeForVoting]
    CreateCategoryRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        activeForVoting:
          type: boolean
    Participant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        categoryId:
          type: integer
        categoryName:
          type: string
        contestantNo:
          type: integer
        photoUrl:
          type: string
          nullable: true
        orderIndex:
          type: integer
      required: [id, name, gender, categoryId, contestantNo, orderIndex]
    CreateParticipantRequest:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        categoryId:
          type: integer
        photoUrl:
          type: string
          nullable: true
      required: [name, gender, categoryId]
    UpdateParticipantRequest:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        categoryId:
          type: integer
        photoUrl:
          type: string
          nullable: true
...